- remove unused functions - remove ruamel_yaml import from active scripts - fix dontcare2background related to PR #256
